+++ docs/make.jl
@@ -6,23 +6,23 @@ using Plots # to not capture precompilation output
 ENV["GKSwstype"] = "100"
 
 # generate the example notebooks for the documentation
-OUTPUT = joinpath(@__DIR__, "..", "docs/examples")
+OUTPUT = joinpath(@__DIR__, "..", "src/generated")
 
-Literate.markdown(joinpath(@__DIR__, "..", "docs/src/bickley.jl"), OUTPUT)
-Literate.notebook(joinpath(@__DIR__, "..", "docs/src/bickley.jl"), OUTPUT; execute=false)
-Literate.script(joinpath(@__DIR__, "..", "docs/src/bickley.jl"), OUTPUT)
+Literate.markdown(joinpath(@__DIR__, "..", "examples/bickley.jl"), OUTPUT)
+Literate.notebook(joinpath(@__DIR__, "..", "examples/bickley.jl"), OUTPUT; execute=false)
+Literate.script(joinpath(@__DIR__, "..", "examples/bickley.jl"), OUTPUT)
 
-Literate.markdown(joinpath(@__DIR__, "..", "docs/src/ocean_flow.jl"), OUTPUT)
-Literate.notebook(joinpath(@__DIR__, "..", "docs/src/ocean_flow.jl"), OUTPUT; execute=false)
-Literate.script(joinpath(@__DIR__, "..", "docs/src/ocean_flow.jl"), OUTPUT)
+Literate.markdown(joinpath(@__DIR__, "..", "examples/ocean_flow.jl"), OUTPUT)
+Literate.notebook(joinpath(@__DIR__, "..", "examples/ocean_flow.jl"), OUTPUT; execute=false)
+Literate.script(joinpath(@__DIR__, "..", "examples/ocean_flow.jl"), OUTPUT)
 
-Literate.markdown(joinpath(@__DIR__, "..", "docs/src/rot_double_gyre.jl"), OUTPUT)
-Literate.notebook(joinpath(@__DIR__, "..", "docs/src/rot_double_gyre.jl"), OUTPUT; execute=false)
-Literate.script(joinpath(@__DIR__, "..", "docs/src/rot_double_gyre.jl"), OUTPUT)
+Literate.markdown(joinpath(@__DIR__, "..", "examples/rot_double_gyre.jl"), OUTPUT)
+Literate.notebook(joinpath(@__DIR__, "..", "examples/rot_double_gyre.jl"), OUTPUT; execute=false)
+Literate.script(joinpath(@__DIR__, "..", "examples/rot_double_gyre.jl"), OUTPUT)
 
-Literate.markdown(joinpath(@__DIR__, "..", "docs/src/standard_map.jl"), OUTPUT)
-Literate.notebook(joinpath(@__DIR__, "..", "docs/src/standard_map.jl"), OUTPUT; execute=false)
-Literate.script(joinpath(@__DIR__, "..", "docs/src/standard_map.jl"), OUTPUT)
+Literate.markdown(joinpath(@__DIR__, "..", "examples/standard_map.jl"), OUTPUT)
+Literate.notebook(joinpath(@__DIR__, "..", "examples/standard_map.jl"), OUTPUT; execute=false)
+Literate.script(joinpath(@__DIR__, "..", "examples/standard_map.jl"), OUTPUT)
 
 # replace links (if any)
 # travis_tag = get(ENV, "TRAVIS_TAG", "")
@@ -44,10 +44,10 @@ makedocs(
     pages = Any[
         "Home" => "index.md"
         "Examples" => [
-            "Rotating Double Gyre" => "rot_double_gyre.md"
-            "Geostrophic Ocean Flow" => "ocean_flow.md"
-            "Bickley Jet" => "bickley.md"
-            "Standard Map" => "standard_map.md"
+            "Rotating double gyre" => "generated/rot_double_gyre.md"
+            "Standard map" => "generated/standard_map.md"
+            "Bickley jet" => "generated/bickley.md"
+            "Geostrophic ocean flow" => "generated/ocean_flow.md"
             ]
         "Basics" => "basics.md"
         "Methods" => [
